Fixing pip installation
We were getting the error `ImportError: cannot import name main` when
trying to install dependencies with pip. According to [the following
link](https://github.com/pypa/pip/issues/5447), it seems that it may be
a problem with pip itself.

Running one of the comments with more upvotes to see if that fixes the
problem