Implementing Dynamo.set_status operation
Implementing an operation to set the status of a specific user. The
method will return True if the status was set successfully, and False if
there was an exception during the process.

The response returned by DynamoDB.put_item does not let us obtain any
information about the item that we just inserted/updated, so the way if
there has been an error during the process is checking that the process
didn't raised any exception. For future iterations, we'll evaluate the
possibility of using the HTTP code included in the Response metadata