Using pytest fixtures and parametrization to mock state and messages
Refactoring actions tests to avoid unnecessary duplicated code. Using
fixtures and indirect test parametrization. More information can be
found on the following links:
* http://docs.pytest.org/en/latest/example/parametrize.html#apply-indirect-on-particular-arguments
* https://stackoverflow.com/questions/18011902/pass-a-parameter-to-a-fixture-function

Also, using `yield` in one of the fixtures to be sure that the patch is
closed after the test ends. See https://pytest.readthedocs.io/en/2.9.1/yieldfixture.html. We will be returning the instance of the mocked state manager, so we can assert if any of its method was called in case we need it