Testing handle in Insulin actions
Adding tests for the `handle` method of the currently existing insulin
actions. On them, we test that the corresponding state is set and that we
send a Telegram reply.

We have also added two new mocks: one to represent a Telegram message
with custom text (included in `conftest.py`) and one to represent the
instance of TelegramWrapper that is imported in `insulin_actions.py`,
used to assert that we're sending the replies