Bug fixing - Unify table schemas for get and put state
We were using different schemas for put and get state operation - On
get we expected the state to be in a column called "value", while on put
we set it in a column named "state". This was caused because "value" was
the value that was planned to be used initially, but later changed to
"state" (as it is more expressive) when implementing the put operation.
However, we forgot to change the get operation.

As we only have unit tests (i.e. we only test either get or put, not
both) we didn't catch this problem before