Encapsulating functional tests helpers and fixtures
Creating a new fixture file, tests.fixtures.functional, in which we'll
store all the functional mocked classes. This will let us have them
centralized in a single place and enable quick interaction among them.

We have also created a helper method to automatically create the
DynamoDB tables for a specific mocked class, and to make it reusable we
create the tables directly on the fixture method before returning the
DynamoDB client. Therefore, all the classes that use it won't need to
think in creating them.

Finally, we have moved the functional state_manager test to a folder
that matches the directory structure of the source code files