Injecting dependencies in actions
Instead of directly instantiating the StateManager and TelegramWrapper
classes in the action initialization, we'll inject them via the
constructor. This allows us to move the responsibility of its creation
to the ActionFactory we added in a previous commit

We have also adapted our tests to this approach. We have created a
subclass of ActionFactory that mocks the dependencies of the actions, in
order to use it during our tests. This allows us to avoid having to mock
all the dependencies in each of the classes where we use an action