Simplifying actions factory + handler tests
Refactoring Actions factory. Now we will just proxy the arguments to the
class constructor unless we specifically implement a different
factory. This will allow us to greatly reduce the amount of code
required for most of the actions

We also added smoke tests for the handler. Their goal is only to check
that the base logic can be executed correctly, without raising any
exceptions.