Fixing permissions + Fixing handler + Deleting commands
Combining several fixes:

1) S3 permissions: when adding DexcomAction, we didn't modify the IAM
user to grant permissions to S3 to upload Dexcom data. Fixing this by
adding the data on `serverless.yml`

2) Fixing handler: The last commit broke the main app logic.
CommandRouter was checking the message text to see if it started with a
'/', but when we send a file there is no message text. We have now
modified the `message_text` property in BaseAction to return an empty
string if there is no text.

3) Deleting commands: as we are not using commands anymore, we'll delete
them to avoid having unnecessary code in the app.