Enabling Meal creation
Adding the possibility to log meals. Currently we'll make it simple and
only allow to add food details, not any of the other fields.

The serializer, SNS client and model were already created from the
previous version of the app so we just needed to create the actions and
its tests