Allowing Insulin doses of negative units
As there is no way of editing an insulin dose once it has been
published, if we incorrectly specify the number of units there is no way
to correct it.

To solve that problem, we'll allow the introduction of insulin doses of
negative units. The idea is that they will be used to compensate for a
registered dose of more units than the ones actually injected (example:
if we get a dose of 4 units but register 44 by mistake, we will just
need to add another dose of -40 units to make it even)