Stop sending SNS message UUID after registering an Injection
The initial idea of the message was to send a follow-up notification that we
could use as the confirmation that the SNS message was published
correctly. However, this reply adds now unnecessary noise, and we have
additional methods to know if injections are correctly being registered
(e.g. the daily basal alarm in alexgascon-api).

This commit stops replying in Telegram with the SNS message ID, and
instead logs it so we can still have it instead it's ever useful